+++ types.h
@@ -346,7 +346,7 @@ Toy_Type*	list_next(Toy_Type *list);
 Toy_Type*	list_get_item(Toy_Type *list);
 int		list_length(Toy_Type *list);
 Toy_Type*	new_integer(mpz_t biginteger);
-Toy_Type*	new_integer_si(int integer);
+Toy_Type*	new_integer_si(long int integer);
 Toy_Type*	new_integer_d(double val);
 char*		integer_to_str(Toy_Type *val);
 Toy_Type*	new_real(double real);
